agents/financial-advisor/pyproject.toml
[project]
name = "financial-advisor"
version = "0.1.0"
description = "AI-driven agent designed to facilitate the exploration of the financial advisor landscape"
authors = [
    { name = "Antonio Gulli", email = "gulli@google.com" }
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "google-adk>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "black>=25.1.0",
    "google-adk[eval]>=1.0.0",
    "pytest-asyncio>=0.26.0",
]
deployment = [
    "absl-py>=2.2.1",
]

[build-system]
requires = ["setuptools>=60", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["financial_advisor"]